pyproject.toml
[tool.poetry]
name = "seamless-re"
version = "0.1.0"
description = ""
authors = ["jdelorimier <delorimier16@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
nltk = "^3.6.5"
opennre = { git = "https://github.com/thunlp/OpenNRE.git" }
torch = "1.6.0"
urllib3 = "1.26.6"
transformers = "3.5.1"
scikit-learn = "^1.0.1"
pandas = "^1.3.4"
ipykernel = "^6.5.1"
spacy = "2.1.0"
python-dotenv = "^0.19.2"
neuralcoref = "^4.0"
smart-open = "^5.2.1"
bs4 = "^0.0.1"
sec-edgar-downloader = "^4.2.2"
flask = "^1.1.2"
pymgclient = "1.1.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
